# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lykke.day",
    "description": "API for managing calendar events",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tasks/today/": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List Todays Tasks",
        "description": "Get all tasks for today.",
        "operationId": "list_todays_tasks_tasks_today__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskSchema"
                  },
                  "type": "array",
                  "title": "Response List Todays Tasks Tasks Today  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/tasks/{date}/{_id}/actions": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Add Task Action",
        "description": "Record an action on a task.",
        "operationId": "add_task_action_tasks__date____id__actions_post",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          },
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": " Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Action"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/task-definitions/{uuid}": {
      "get": {
        "tags": [
          "task-definitions"
        ],
        "summary": "Get Task Definition",
        "description": "Get a single task definition by ID.",
        "operationId": "get_task_definition_task_definitions__uuid__get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDefinitionSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "task-definitions"
        ],
        "summary": "Update Task Definition",
        "description": "Update a task definition.",
        "operationId": "update_task_definition_task_definitions__uuid__put",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskDefinitionUpdateSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDefinitionSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "task-definitions"
        ],
        "summary": "Delete Task Definition",
        "description": "Delete a task definition.",
        "operationId": "delete_task_definition_task_definitions__uuid__delete",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/task-definitions/": {
      "get": {
        "tags": [
          "task-definitions"
        ],
        "summary": "List Task Definitions",
        "description": "List task definitions with pagination.",
        "operationId": "list_task_definitions_task_definitions__get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskDefinitionSchema"
                  },
                  "title": "Response List Task Definitions Task Definitions  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "task-definitions"
        ],
        "summary": "Create Task Definition",
        "description": "Create a new task definition.",
        "operationId": "create_task_definition_task_definitions__post",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskDefinitionCreateSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDefinitionSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/task-definitions/bulk/": {
      "post": {
        "tags": [
          "task-definitions"
        ],
        "summary": "Bulk Create Task Definitions",
        "description": "Bulk create task definitions.",
        "operationId": "bulk_create_task_definitions_task_definitions_bulk__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array",
                "title": "Task Definitions Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskDefinitionSchema"
                  },
                  "type": "array",
                  "title": "Response Bulk Create Task Definitions Task Definitions Bulk  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
...
    "/routines/{uuid}/tasks": {
      "post": {
        "tags": [
          "routines"
        ],
        "summary": "Add Routine Task",
        "description": "Attach a task definition to a routine.",
        "operationId": "add_routine_task_routines__uuid__tasks_post",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutineTaskCreateSchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/routines/{uuid}/tasks/{task_definition_id}": {
      "put": {
        "tags": [
          "routines"
        ],
        "summary": "Update Routine Task",
        "description": "Update an attached routine task (name/schedule).",
        "operationId": "update_routine_task_routines__uuid__tasks__task_definition_id__put",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Uuid"
            }
          },
          {
            "name": "task_definition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Definition Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutineTaskUpdateSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "routines"
        ],
        "summary": "Remove Routine Task",
        "description": "Detach a task definition from a routine.",
        "operationId": "remove_routine_task_routines__uuid__tasks__task_definition_id__delete",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Uuid"
            }
          },
          {
            "name": "task_definition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Definition Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Action": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/ActionType"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "Action"
      },
      "ActionSchema": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/ActionType"
          },
          "data": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Data"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "type",
          "created_at"
        ],
        "title": "ActionSchema",
        "description": "API schema for Action value object."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoutineSchema": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "$ref": "#/components/schemas/TaskCategory"
          },
          "routine_schedule": {
            "$ref": "#/components/schemas/RoutineSchedule"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/RoutineTask-Output"
            },
            "type": "array",
            "title": "Tasks"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "name",
          "category",
          "routine_schedule"
        ],
        "title": "RoutineSchema",
        "description": "API schema for Routine entity."
      },
...
      "RoutineTaskCreateSchema": {
        "properties": {
          "task_definition_id": {
            "type": "string",
            "format": "uuid",
            "title": "Task Definition Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "schedule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskSchedule"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "task_definition_id"
        ],
        "title": "RoutineTaskCreateSchema",
        "description": "API schema for attaching a RoutineTask."
      },
      "RoutineTaskUpdateSchema": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "schedule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskSchedule"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "RoutineTaskUpdateSchema",
        "description": "API schema for updating an attached RoutineTask."
      },
...
      "TaskCategory": {
        "type": "string",
        "enum": [
          "HYGIENE",
          "GROOMING",
          "SLEEP",
          "HEALTH",
          "FITNESS",
          "MEDICAL",
          "NUTRITION",
          "COOKING",
          "HOUSE",
          "CLEANING",
          "MAINTENANCE",
          "ORGANIZATION",
          "WORK",
          "MEETING",
          "PROFESSIONAL",
          "FAMILY",
          "SOCIAL",
          "RELATIONSHIP",
          "SHOPPING",
          "ERRAND",
          "COMMUTE",
          "TRAVEL",
          "ENTERTAINMENT",
          "HOBBY",
          "RECREATION",
          "EDUCATION",
          "LEARNING",
          "FINANCE",
          "BILLS",
          "TECHNOLOGY",
          "DIGITAL",
          "PET",
          "PLANNING",
          "ADMIN"
        ],
        "title": "TaskCategory"
      },
      "TaskDefinitionCreateSchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "type": {
            "$ref": "#/components/schemas/TaskType"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "type"
        ],
        "title": "TaskDefinitionCreateSchema",
        "description": "API schema for creating a TaskDefinition entity."
      },
      "TaskDefinitionSchema": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "type",
          "user_id"
        ],
        "title": "TaskDefinitionSchema",
        "description": "API schema for TaskDefinition entity."
      },
      "TaskDefinitionUpdateSchema": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskType"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "TaskDefinitionUpdateSchema",
        "description": "API schema for TaskDefinition update requests."
      },
      "TaskFrequency": {
        "type": "string",
        "enum": [
          "DAILY",
          "CUSTOM_WEEKLY",
          "WEEKLY",
          "ONCE",
          "YEARLY",
          "MONTHLY",
          "BI_WEEKLY",
          "WORK_DAYS",
          "WEEKENDS"
        ],
        "title": "TaskFrequency"
      },
...
      "TaskScheduleSchema": {
        "properties": {
          "available_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Time"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "timing_type": {
            "$ref": "#/components/schemas/TimingType"
          }
        },
        "type": "object",
        "required": [
          "timing_type"
        ],
        "title": "TaskScheduleSchema",
        "description": "API schema for TaskSchedule value object."
      },
      "TaskSchema": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "scheduled_date": {
            "type": "string",
            "format": "date",
            "title": "Scheduled Date"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "task_definition": {
            "$ref": "#/components/schemas/TaskDefinitionSchema"
          },
          "category": {
            "$ref": "#/components/schemas/TaskCategory"
          },
          "frequency": {
            "$ref": "#/components/schemas/TaskFrequency"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "schedule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskScheduleSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "routine_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Routine Id"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TaskTag"
            },
            "type": "array",
            "title": "Tags"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionSchema"
            },
            "type": "array",
            "title": "Actions"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "scheduled_date",
          "name",
          "status",
          "task_definition",
          "category",
          "frequency"
        ],
        "title": "TaskSchema",
        "description": "API schema for Task entity."
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "COMPLETE",
          "NOT_READY",
          "READY",
          "PUNT",
          "NOT_STARTED",
          "PENDING"
        ],
        "title": "TaskStatus"
      },
      "TaskTag": {
        "type": "string",
        "enum": [
          "AVOIDANT",
          "FORGETTABLE",
          "IMPORTANT",
          "URGENT",
          "FUN"
        ],
        "title": "TaskTag"
      },
...
      "TimingType": {
        "type": "string",
        "enum": [
          "DEADLINE",
          "FIXED_TIME",
          "TIME_WINDOW",
          "FLEXIBLE"
        ],
        "title": "TimingType"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "lykke_auth"
      }
    }
  }
}